Added a Steps placeholder page, hardcoded step url from tasks, moved around some routes.
