+++ src/frontend/App.js
@@ -2,12 +2,11 @@ import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
 import Header from './components/Header';
-import ModuleHome from './components/Module/ModuleHome';
 import About from './pages/About';
-import Module from './components/Module/Modules';
-import Phase from './components/Phase/Phases';
-// import Task from './components/Task/Tasks';
-// import Step from './components/Step/Step';
+import Modules from './components/Module/Modules';
+import Phases from './components/Phase/Phases';
+import Tasks from './components/Task/Tasks';
+import Steps from './components/Step/Steps';
 // import Guidance from './components/Guidance/Guidances';
 import PageNotFound from './pages/PageNotFound';
 import '@gctools-components/aurora-ds/css/aurora.min.css';
@@ -33,14 +32,14 @@ export default class App extends Component {
         <main id="main-content" role="main">
           <div className="container">
             <Switch>
-              <Route exact path="/" component={ModuleHome} />
-              <Route exact path="/modules/" component={ModuleHome} />
+              <Route exact path="/" component={Modules} />
+              <Route exact path="/modules/" component={Modules} />
               <Route exact path="/about" component={About} />
-              <Route exact path="/modules/:moduleId/phases/" component={Module} />
-              <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/" component={Phase} />
-              {/* <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/:taskId/steps/" component={Task} />
-              <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/:taskId/steps/:stepId/guidances/" component={Step} />
-              <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/:taskId/steps/:stepId/guidances/:guidanceId" component={Guidance} /> */}
+              <Route exact path="/modules/:moduleId/phases/" component={Phases} />
+              <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/" component={Tasks} />
+              <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/:taskId/steps/:stepId" component={Steps} />
+              {/* <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/:taskId/steps/:stepId/guidances/" component={Step} /> */}
+              {/* <Route exact path="/modules/:moduleId/phases/:phaseId/tasks/:taskId/steps/:stepId/guidances/:guidanceId" component={Guidance} /> */}
               <Route component={PageNotFound} />
             </Switch>
           </div>
