--- src/frontend/components/Step/Step.js
@@ -1,5 +1,0 @@
-/* Todo: https://sketch.cloud/s/ng0Yl/a/Qvv39w
-  Step:
-    |_StepIndicator
-    |_StepCard
-*/
