+++ src/frontend/components/Task/Tasks.js
@@ -1,11 +1,10 @@
-/* Todo: https://sketch.cloud/s/ng0Yl/a/KbYEay
+/* @todo: https://sketch.cloud/s/ng0Yl/a/KbYEay
   Tasks:
     |_Parent Phase Title
     |_Title Text: TaskList
     |_TaskList
       |_TaskCard
 */
-
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import TaskList from './TaskList';
@@ -13,20 +12,17 @@ import TaskList from './TaskList';
 export default class Tasks extends Component {
   constructor() {
     super();
-
     this.state = {
-      phase: null,
-      tasks: []
+      phase: null
     };
   }
 
-  // Fetch the list on first mount
   componentDidMount() {
     const { match } = this.props;
-    const { params } = match;
-    const { phaseId } = params;
-    this.getPhasebyId(phaseId);
-    this.getTasks(phaseId);
+    const { phaseId } = match.params;
+    if (phaseId) {
+      this.getPhasebyId(phaseId);
+    }
   }
 
   getPhasebyId = (phaseId) => {
@@ -40,39 +36,32 @@ export default class Tasks extends Component {
       });
   }
 
-  // Retrieves the list of items from the Express app
-  getTasks = (phaseId) => {
-    fetch(`/api/tasks?phase_id=${phaseId}`)
-      .then(res => res.json())
-      .then(tasks => this.setState({ tasks }));
-  };
-
   render() {
-    const { tasks, phase } = this.state;
-    return (
-      <div>
-        <DisplayPhaseInfo module={phase} />
-        <TaskList tasks={tasks} />
-      </div>
-    );
+    const { phase } = this.state;
+    if (phase !== null) {
+      const { tasks } = phase;
+      return (
+        <div>
+          <DisplayPhaseInfo phase={phase} />
+          {(tasks) ? <TaskList tasks={tasks} /> : ''}
+        </div>
+      );
+    }
+    return '';
   }
 }
 
 function DisplayPhaseInfo({ phase }) {
-  if (phase !== null) {
-    return (
-      <div>
-        <h1>
-          {`Phase ${phase.orderNum}:`}
-          {phase.titleEng}
-        </h1>
-        <h2>
-          {'Task List'}
-        </h2>
-      </div>
-    );
-  }
-  return '';
+  return (
+    <div>
+      <h1 className="h6">
+        {`Phase ${phase.orderNum}: ${phase.titleEng}`}
+      </h1>
+      <h2 className="h2">
+        {'Task List'}
+      </h2>
+    </div>
+  );
 }
 
 Tasks.defaultProps = {
@@ -91,4 +80,14 @@ Tasks.propTypes = {
       phaseId: PropTypes.string.isRequired
     })
   })
+};
+
+DisplayPhaseInfo.propTypes = {
+  phase: PropTypes.shape(
+    {
+      phase_id: PropTypes.string.isRequired,
+      titleEng: PropTypes.string.isRequired,
+      orderNum: PropTypes.string.isRequired,
+    }
+  ).isRequired
 };
